Migrate addCustomerReference to CoreData class
