+++ Helper/Data.php
@@ -48,11 +48,6 @@ use Magento\Store\Api\Data\StoreInterface;
 use Magento\Store\Model\Store;
 use Nosto\Tagging\Helper\Scope as NostoHelperScope;
 use phpseclib\Crypt\Random;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Customer\Model\Customer;
-use Magento\Eav\Model\Entity\Attribute\SetFactory as AttributeSetFactory;
-use Magento\Customer\Setup\CustomerSetupFactory;
-use Magento\Framework\Exception\LocalizedException;
 
 /**
  * NostoHelperData helper used for common tasks, mainly configurations.
@@ -210,8 +205,6 @@ class Data extends AbstractHelper
     private $productMetaData;
     private $nostoHelperScope;
     private $cacheManager;
-    private $customerSetupFactory;
-    private $attributeSetFactory;
 
     /**
      * Data constructor.
@@ -221,8 +214,6 @@ class Data extends AbstractHelper
      * @param WriterInterface $configWriter
      * @param ProductMetadataInterface $productMetadataInterface
      * @param CacheManager $cacheManager
-     * @param CustomerSetupFactory $customerSetupFactory
-     * @param AttributeSetFactory $attributeSetFactory
      */
     public function __construct(
         Context $context,
@@ -230,9 +221,7 @@ class Data extends AbstractHelper
         ModuleListInterface $moduleListing,
         WriterInterface $configWriter,
         ProductMetadataInterface $productMetadataInterface,
-        CacheManager $cacheManager,
-        CustomerSetupFactory $customerSetupFactory,
-        AttributeSetFactory $attributeSetFactory
+        CacheManager $cacheManager
     ) {
         parent::__construct($context);
 
@@ -241,8 +230,6 @@ class Data extends AbstractHelper
         $this->productMetaData = $productMetadataInterface;
         $this->nostoHelperScope = $nostoHelperScope;
         $this->cacheManager = $cacheManager;
-        $this->customerSetupFactory = $customerSetupFactory;
-        $this->attributeSetFactory = $attributeSetFactory;
     }
 
     /**
@@ -667,54 +654,5 @@ class Data extends AbstractHelper
         if (!empty($clearTypes)) {
             $this->cacheManager->clean($clearTypes);
         }
-    }
-
-    /**
-     * @param ModuleDataSetupInterface $setup
-     * @throws LocalizedException
-     * @throws \Zend_Validate_Exception
-     */
-    public function addCustomerReference(ModuleDataSetupInterface $setup)
-    {
-        $customerSetup = $this->customerSetupFactory->create(['setup' => $setup]);
-
-        $customerEntity = $customerSetup->getEavConfig()->getEntityType(Customer::ENTITY);
-        $attributeSetId = (int)$customerEntity->getDefaultAttributeSetId();
-
-        $attributeSet = $this->attributeSetFactory->create();
-        $attributeGroupId = $attributeSet->getDefaultGroupId($attributeSetId);
-
-        $customerSetup->addAttribute(
-            Customer::ENTITY,
-            self::NOSTO_CUSTOMER_REFERENCE_ATTRIBUTE_NAME,
-            [
-                'type' => 'varchar',
-                'label' => 'Nosto Customer Reference',
-                'input' => 'text',
-                'required' => false,
-                'sort_order' => 120,
-                'position' => 120,
-                'visible' => true,
-                'user_defined' => true,
-                'unique' => true,
-                'system' => false,
-            ]
-        );
-
-        $attribute = $customerSetup->getEavConfig()->getAttribute(
-            Customer::ENTITY,
-            self::NOSTO_CUSTOMER_REFERENCE_ATTRIBUTE_NAME
-        );
-
-        $attribute->addData(
-            [
-                'attribute_set_id' => $attributeSetId,
-                'attribute_group_id' => $attributeGroupId,
-                'used_in_forms' => ['adminhtml_customer', 'customer_account_edit'],
-            ]
-        );
-
-        // @codingStandardsIgnoreLine
-        $attribute->save();
     }
 }
